Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit 886e26dd71c4cf4653bbef7b1209677218ceb70d, reversing
changes made to da18f911071d65582f3830a80a7696085f3419a5.

+++ src/diamond/handler/graphite.py
@@ -39,7 +39,7 @@ class GraphiteHandler(Handler):
         self.proto = self.config['proto'].lower().strip()
         self.host = self.config['host']
         self.port = int(self.config['port'])
-        self.timeout = float(self.config['timeout'])
+        self.timeout = int(self.config['timeout'])
         self.keepalive = bool(self.config['keepalive'])
         self.keepaliveinterval = int(self.config['keepaliveinterval'])
         self.batch_size = int(self.config['batch'])
